+++ rimfs/src/fs/ext4/types/mod.rs
@@ -1,13 +1,17 @@
 // SPDX-License-Identifier: MIT
 
 pub mod bgdt;
+pub mod bgdt_update;
 pub mod dirent;
 pub mod extent;
 pub mod inode;
 pub mod superblock;
 
 pub use bgdt::*;
+pub use bgdt_update::*;
 pub use dirent::*;
 pub use extent::*;
 pub use inode::*;
+mod lost_found;
+pub use lost_found::*;
 pub use superblock::*;